Moving renderer-focused workday-waiver-aux.js to renderer folder
